+++ src/crisptrutski/boot_cljs_test/utils.clj
@@ -1,8 +1,11 @@
 (ns crisptrutski.boot-cljs-test.utils
   (:require
+    [boot.core :as boot]
     [boot.pod :as pod]
+    [boot.util :refer [warn]]
+    [clojure.java.io :as io]
     [clojure.string :as str]
-    [boot.core :as core])
+    [crisptrutski.boot-error.core :as err])
   (:import
     [java.io File]
     [java.util.regex Pattern]
@@ -28,33 +31,20 @@
       (str/replace #"\/|\\" ".")
       symbol))
 
-(defn ns-literal? [ns]
-  (symbol? ns))
-
 (defn ns-regex [ns]
-  (cond
-    (instance? Pattern ns) ns
-    :else (re-pattern (str "\\A" (name ns) "\\z"))))
-
-(defn ns->cljs-path
-  "Determine filename from namespace"
-  [ns]
-  (-> (str ns)
-      (str/replace "-" "_")
-      (str/replace "." "/")
-      (str ".cljs")))
+  (if (instance? Pattern ns) ns (re-pattern (str "\\A" (name ns) "\\z"))))
 
 (defn ^Path filename->path
   "A sane constructor `java.nio.file.Path` instances."
   [filename]
-  ;; Workaround for bad time with types (can't hint (Paths/get ^String))
+  ;; Workaround for bad time with native dispatch and Java splat.
+  ;; Otherwise dispatches to URI even with ^String hint.
   (Paths/get filename (into-array String nil)))
 
 (defn relativize
   "Get the relative path from `a` to `b`"
   [a b]
-  (str (.relativize (filename->path a)
-                    (filename->path b))))
+  (str (.relativize (filename->path a) (filename->path b))))
 
 (defn src-file?
   "Is the file a Clojure(Script) file?"
@@ -74,21 +64,47 @@
        (filter src-file?)
        (map #(file->ns (relativize dir %)))))
 
-(defn refine-namespaces [fs namespaces]
-  (if (and (seq namespaces) (every? ns-literal? namespaces))
+(defn refine-namespaces [fs namespaces exclusions]
+  (if (and (seq namespaces) (every? symbol? namespaces))
     namespaces
-    (let [regexes (map ns-regex namespaces)]
-      (->> (core/input-dirs fs)
-           (mapv (memfn getPath))
+    (let [regexes (map ns-regex namespaces)
+          exclude (map ns-regex exclusions)]
+      (->> (boot/input-dirs fs)
+           (mapv #(.getPath %))
            (into #{} (mapcat ns-from-dir))
-           (filter (fn [ns] (or (and (empty? regexes)
-                                     (re-find #"-test\z" (str ns)))
-                                (some #(re-find % (str ns)) regexes))))))))
+           (filter (fn [ns]
+                     (let [s (str ns)]
+                       (and (or (and (empty? regexes) (re-find #"-test\z" s))
+                                (some #(re-find % s) regexes))
+                            (not (some #(re-find % s) exclude))))))))))
 
 (defn cljs-files
   "Given a fileset, return a list of all the ClojureScript source files."
   [fileset]
   (->> fileset
-       core/input-files
-       (core/by-ext [".cljs" ".cljc"])
+       boot/input-files
+       (boot/by-ext [".cljs" ".cljc" ".cljs.edn"])
        (sort-by :path)))
+
+(defn find-path
+  "Fine most recent filesystem file in fileset with given logical path."
+  [fileset filename]
+  (some->>
+    (boot/output-files fileset)
+    (filter (comp #{filename} :path))
+    (sort-by :time)
+    last ^File boot/tmp-file .getPath))
+
+(defn build-cljs-opts
+  "Augment cljs-opts with boot determined values"
+  [base output-to output-dir]
+  (merge {:asset-path (.getName (io/file output-dir))}
+         base
+         {:output-to output-to :output-dir output-dir}))
+
+(defn combine-cljs-opts
+  "Augment cljs-opts with values determined by other arguments"
+  [cljs-opts optimizations js-env]
+  (merge {:optimizations optimizations}
+         (when (= :node js-env) {:target :nodejs, :hashbang false})
+         cljs-opts))